動的テンプレート変換サービスインタフェースと入出力サービスインタフェースを追加し、関連するサービスのインポートを修正
