+++ src/main/java/kmg/tool/application/service/io/AccessorCreationService.java
@@ -1,6 +1,6 @@
 package kmg.tool.application.service.io;
 
-import kmg.tool.domain.service.Two2OneService;
+import kmg.tool.domain.service.io.Two2OneService;
 
 /**
  * アクセサ作成サービスインタフェース<br>
