+++ src/main/java/kmg/tool/application/service/io/SimpleTwo2OneService.java
@@ -1,6 +1,6 @@
 package kmg.tool.application.service.io;
 
-import kmg.tool.domain.service.Two2OneService;
+import kmg.tool.domain.service.io.Two2OneService;
 
 /**
  * シンプル2入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
