+++ src/main/java/kmg/tool/domain/service/io/Two2OneService.java
@@ -1,13 +1,13 @@
-package kmg.tool.domain.service;
+package kmg.tool.domain.service.io;
 
 import java.nio.file.Path;
 
 import kmg.tool.infrastructure.exception.KmgToolException;
 
 /**
- * 1入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
+ * シンプル2入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
  */
-public interface One2OneService extends IoService {
+public interface Two2OneService extends IoService {
 
     /**
      * 入力ファイルパスを返す<br>
@@ -36,17 +36,32 @@ public interface One2OneService extends IoService {
     Path getOutputPath();
 
     /**
+     * テンプレートファイルパスを返す<br>
+     *
+     * @author KenichiroArai
+     *
+     * @sine 1.0.0
+     *
+     * @version 1.0.0
+     *
+     * @return テンプレートファイルパス
+     */
+    Path getTemplatePath();
+
+    /**
      * 初期化する
      *
      * @return true：成功、false：失敗
      *
      * @param inputPath
-     *                   入力ファイルパス
+     *                     入力ファイルパス
+     * @param templatePath
+     *                     テンプレートファイルパス
      * @param outputPath
-     *                   出力ファイルパス
+     *                     出力ファイルパス
      *
      * @throws KmgToolException
      *                          KMGツール例外
      */
-    boolean initialize(Path inputPath, Path outputPath) throws KmgToolException;
+    boolean initialize(final Path inputPath, final Path templatePath, final Path outputPath) throws KmgToolException;
 }
