+++ src/main/java/kmg/tool/presentation/ui/cli/io/AbstractTwo2OneTool.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import kmg.core.infrastructure.utils.KmgPathUtils;
 import kmg.foundation.infrastructure.context.KmgMessageSource;
-import kmg.tool.domain.service.Two2OneService;
+import kmg.tool.domain.service.io.Two2OneService;
 import kmg.tool.domain.types.KmgToolLogMessageTypes;
 import kmg.tool.infrastructure.exception.KmgToolException;
